requirements.txt
Django==2.2.8
django-jet==1.0.8
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
mysqlclient==1.4.6
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
